==> notebooks/01_feature_importance.ipynb <==
# 1. Info

This notebooks contains the feature importance analysis for the hotel cancelation dataset.

Before runing this notebook you should have run the "01_data_preparation.ipynb" notebook.

# 2. Feature importance analysis.

## 2.1. Import Libraries

import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.metrics import mutual_info_score
from IPython.display import display

## 2.2. Read the data

data = pd.read_csv('../data/Hotel_Cancelations.csv')

data.drop(columns=['Booking_ID'], axis=1, inplace=True)

## 2.3. Setting up the validation framework

categorical_variables = ['type_of_meal_plan','room_type_reserved','market_segment_type']
numerical_variables = ['no_of_adults','no_of_children','no_of_weekend_nights','no_of_week_nights','required_car_parking_space','lead_time','arrival_year','arrival_month','arrival_date','repeated_guest',
'no_of_previous_cancellations','no_of_previous_bookings_not_canceled','avg_price_per_room','no_of_special_requests']
target_variable = ['booking_status']


df_full_train, df_test = train_test_split(data, test_size=0.2, random_state=1)
df_train, df_val = train_test_split(df_full_train, test_size=0.25, random_state=1)

print(len(df_train))
print(len(df_test))
print(len(df_val))

df_train = df_train.reset_index(drop=True)
df_val = df_val.reset_index(drop=True)
df_test = df_test.reset_index(drop=True)

y_train = df_train.booking_status.values
y_val = df_val.booking_status.values
y_test = df_test.booking_status.values

del df_train['booking_status']
del df_val['booking_status']
del df_test['booking_status']

## 2.4 difference and risk ratio

global_cancelation = data.booking_status.mean()

for cat in categorical_variables:
    print(cat)
    df_group = data.groupby(cat).booking_status.agg(['mean', 'count'])
    df_group['diff'] = df_group['mean'] - global_cancelation
    df_group['risk_ratio'] = df_group['mean'] / global_cancelation
    display(df_group)
    print()
    print()

## 2.5. Mutual information (Categorical variables)

def mutual_info_cancelation_score(series):
    return mutual_info_score(series, df_full_train.booking_status)

mi = df_full_train[categorical_variables].apply(mutual_info_cancelation_score)
mi.sort_values(ascending=False)

The only categorical feature that will be considered for the model selection is the market_segment_type.

## 2.6 Correlation (Numerical variables)

numerical_corr = df_full_train[numerical_variables].corrwith(df_full_train.booking_status).abs().reset_index()
numerical_corr.columns = ['features','corr']
numerical_corr.sort_values(by=['corr'], ascending=False)

numerical_corr.query(f"corr > {numerical_corr['corr'].mean()}")['features'].values

lead_time, arrival_year, avg_price_per_room, and no_of_special_requests are the numerical features that will be considered for the model selection.

End of notebook.
