==> notebooks/04_predict.ipynb <==
# 1. Info

Notebook to make predictions using the **xgboost_model.pkl** and **dv.pkl** files.

Before running this notebook run the notebook "03_train_model.ipynb".

# 2. Make predictions

import pickle
import xgboost as xgb

## 2.1. Read model

# files path
model_file = '../xgboost_model.pkl'
dv_file = '../dv.pkl'

# read the model
with open(model_file, 'rb') as f_in:
    model = pickle.load(f_in)

# read the dv
with open(dv_file, 'rb') as f_in:
    dv = pickle.load(f_in)

## 2.2. Data to predict

data = {
    'lead_time':39,
    'arrival_year':2019,
    'avg_price_per_room':102.12,
    'no_of_special_requests':2,
    'market_segment_type':1,
}

## 2.3. Vectorize the data

X = dv.fit_transform(data)
features = dv.feature_names_
dX = xgb.DMatrix(X, feature_names=features)

## 2.4. Make predictions

y_pred = model.predict(dX)[0]
bool(y_pred >= 0.5)

y_pred

End of the notebook
