==> notebooks/05_test_services.ipynb <==
# 1. Info

Notebook to test the services. Local and cloud.

# 2. Test local service

import requests

data = {
    'lead_time':39,
    'arrival_year':2019,
    'avg_price_per_room':102.12,
    'no_of_special_requests':2,
    'market_segment_type':1,
}

local_url = 'http://0.0.0.0:9696/hotel_cancelation'

requests.post(local_url, json=data).json()

# 3. Test cloud service

import requests

data = {
    'lead_time':39,
    'arrival_year':2019,
    'avg_price_per_room':102.12,
    'no_of_special_requests':2,
    'market_segment_type':1,
}

cloud_url = 'https://hotel-cancelation.onrender.com/hotel_cancelation'

requests.post(cloud_url, json=data).json()

End of the notebook
